Refactor: Publish Page 컴포넌트 나누기
